# Conflicts:
#	src/main/java/com/keymanager/monitoring/service/CustomerChargeRuleTypeService.java
#	src/main/webapp/customer/customerlist.jsp

+++ src/main/java/com/keymanager/monitoring/dao/CustomerChargeTypeDao.java
@@ -8,6 +8,6 @@ public interface CustomerChargeTypeDao extends BaseMapper<CustomerChargeType> {
 
     int selectLastId();
 
-    CustomerChargeType selectByCustomerUuid(@Param("customerUuid") Long customerUuid);
+    CustomerChargeType selectByCustomerUuid(@Param("customerUuid") Integer customerUuid);
 
 }
